Attempts to make login functions
It's not working but now I know why. It's because the variables inside
the function being called (like get_post_inputs()) are not able to be
referenced outside of it (as in the login() function). I think this can
be resolved by making it a class. Also note that the reason $_POST
worked okay was because it is a global variable.